feat: add subcommand support
